Report unhealthy if the hub kubeconfig certificate changed
This will cause the deployment to restart if the certificate changed but
the hub kubeconfig did not. This happens since the kubeconfig only
specifies a path to the certificate, which doesn't change if the
certificate is renewed.

Relates:
https://issues.redhat.com/browse/ACM-10500

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 050bf23e696209c4a80612cf8763f82cd2710a42)