Handle hub template errors
The plan is to add template support to more policy types. Currently it
is restricted to ConfigurationPolicies. This removes that restriction
and reports hub template errors that would be added by the propagator,
so that each policy controller does not need to duplicate the logic to
emit those events themselves.

Refs:
 - https://issues.redhat.com/browse/ACM-10858

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 4d7c4dadf3630e089466698b952cf91b66b130e3)