Adjust template-sync event filter
The compliance events emitted by the template-sync controller should
have priority over any created by the policy controllers. For example,
in the past we had issues where a "late" compliance event from a
controller would override the Pending status of a policy, and there are
other template-error statuses that are emitted by the template-sync
which could have the same situation. To ensure that a template-sync event
is always emitted after a different compliance event, this controller
should reconcile in more situations.

Note that it might seem better if the template-sync controller removed
the templates that have had an error, but in the case of
ConfigurationPolicy, removing it can cause undesirable side-effects due
to PruneObjectBehavior.

Refs:
 - https://issues.redhat.com/browse/ACM-10858

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 311debf5fce423a55eca7473c93a89bde32ed1d6)